Moved more controls to the sidebar
