+++ web_src/main.js
@@ -210,81 +210,6 @@ async function loadMap(mapParam) {
           }),
         },
       }),
-      // share button
-      L.Toolbar2.Action.extend({
-        options: {
-          toolbarIcon: { html: '<i class="fa-solid fa-link"></i>', tooltip: 'Share' },
-          subToolbar: new L.Toolbar2({
-            actions: [
-              subAction.extend({
-                options: { toolbarIcon: { html: 'Copy Map View URL', tooltip: 'Copies View URL to the Clipboard' } },
-                addHooks: function () {
-                  browser.copyTextToClipboard(MapParam.getViewURL(map));
-                  subAction.prototype.addHooks.call(this); // closes sub-action
-                },
-              }),
-              subAction.extend({
-                options: { toolbarIcon: { html: '&times;', tooltip: 'Close' } },
-              }),
-            ],
-          }),
-        },
-      }),
-      // load game button
-      L.Toolbar2.Action.extend({
-        options: {
-          toolbarIcon: { html: '<i class="fa-regular fa-folder-open"></i>', tooltip: 'Browse...' },
-          subToolbar: new L.Toolbar2({
-            actions: [
-              subAction.extend({
-                options: {
-                  toolbarIcon: { html: 'Browse...', tooltip: 'Load game save (*.sav) to mark collected items (Alt+R)' },
-                },
-                addHooks: function () {
-                  if (
-                    Object.keys(Settings.map.saveData).length == 0 ||
-                    skipConfirms ||
-                    confirm('Are you sure you want to overwrite existing items marked found?')
-                  ) {
-                    SaveFileSystem.loadFileDialog(map.mapId);
-                  }
-                  subAction.prototype.addHooks.call(this);
-                },
-              }),
-              subAction.extend({
-                options: {
-                  toolbarIcon: {
-                    html: 'Copy File Path',
-                    tooltip: 'Copy default Windows game save file path to the Clipboard',
-                  },
-                },
-                addHooks: function () {
-                  const savedPaths = {
-                    sl: '%LocalAppData%\\Supraland\\Saved\\SaveGames',
-                    slc: '%LocalAppData%\\Supraland\\Saved\\SaveGames',
-                    siu: '%LocalAppData%\\SupralandSIU\\Saved\\SaveGames',
-                    sw: '%LocalAppData%\\Supraworld\\Saved\\SaveGames\\Supraworld',
-                  };
-                  browser.copyTextToClipboard(savedPaths[map.mapId]);
-                  subAction.prototype.addHooks.call(this);
-                },
-              }),
-              subAction.extend({
-                options: { toolbarIcon: { html: 'Unmark Found', tooltip: 'Unmark all found items' } },
-                addHooks: function () {
-                  if (skipConfirms || confirm('Are you sure to unmark all found items?')) {
-                    SaveFileSystem.ClearAll();
-                  }
-                  subAction.prototype.addHooks.call(this);
-                },
-              }),
-              subAction.extend({
-                options: { toolbarIcon: { html: '&times;', tooltip: 'Close' } },
-              }),
-            ],
-          }),
-        },
-      }),
     ],
   }).addTo(map);
